check if my changes solid
