--- examples/example_plugin/example_plugin/ui/NonModelView.jsx
@@ -1,10 +1,0 @@
-import GenericView from "@views/generic/GenericView";
-
-
-export default function NonModelView(){
-    return (
-        <GenericView gridBackground="white">
-            <div> Im a humble 🙈 Non Model Plugin View.</div>
-        </GenericView>
-    )
-}
